Added jline-console module
