+++ console/src/main/java/org/jline/console/AbstractCommandRegistry.java
