+++ console/src/main/java/org/jline/console/CmdDesc.java
