+++ demo/src/main/java/org/jline/demo/Repl.java
@@ -26,12 +26,16 @@ import java.util.stream.Collectors;
 
 import org.jline.builtins.*;
 import org.jline.builtins.Completers.OptionCompleter;
-import org.jline.builtins.Widgets.TailTipWidgets;
-import org.jline.builtins.Widgets.TailTipWidgets.TipType;
+import org.jline.console.impl.Builtins;
+import org.jline.console.impl.ConsoleEngineImpl;
+import org.jline.console.impl.SystemRegistryImpl;
+import org.jline.console.Widgets.TailTipWidgets;
+import org.jline.console.Widgets.TailTipWidgets.TipType;
 import org.jline.console.CommandInput;
 import org.jline.console.CommandMethods;
 import org.jline.console.CommandRegistry;
-import org.jline.console.ConfigurationPath;
+import org.jline.console.ConsoleEngine;
+import org.jline.console.JlineCommandRegistry;
 import org.jline.keymap.KeyMap;
 import org.jline.reader.*;
 import org.jline.reader.LineReader.Option;
